Refactor team access
